+++ src/server/types/kras-middleware.ts
@@ -2,6 +2,7 @@ import { KrasServerHandler } from './kras-express';
 
 export interface KrasMiddleware {
   source: string;
+  direction: 'in' | 'out';
   options: Array<any>;
   active: boolean;
   handler: KrasServerHandler;
@@ -9,6 +10,7 @@ export interface KrasMiddleware {
 
 export interface KrasMiddlewareDefinition {
   source: string;
+  direction?: 'in' | 'out';
   baseDir?: string;
   options?: Array<any>;
 }